replace unsigned int by uint16_t

+++ PWM_TP/src/configuration.c
@@ -4,7 +4,7 @@
 /*											Timer Configuration													*/
 /*--------------------------------------------------------------------------------------------------------------*/
 
-void Timer_Config(TIM_HandleTypeDef * Timer, TIM_TypeDef * timer_choice, unsigned int prescaler, unsigned int period){
+void Timer_Config(TIM_HandleTypeDef * Timer, TIM_TypeDef * timer_choice, uint16_t prescaler, uint16_t period){
 										
 	ENABLE_CLK_TIMER_CHOICE(timer_choice);
 										
@@ -17,7 +17,7 @@ void Timer_Config(TIM_HandleTypeDef * Timer, TIM_TypeDef * timer_choice, unsigne
 	HAL_TIM_Base_Init(Timer);
 }
 									
-void Timer_Mode_OCPWM_Config(TIM_HandleTypeDef * Timer, unsigned int mode, unsigned int polarity){
+void Timer_Mode_OCPWM_Config(TIM_HandleTypeDef * Timer, uint16_t mode, uint16_t polarity){
 	HAL_TIM_OC_Init(Timer);
 	TIM_OC_InitTypeDef TIM_PWM_OC;
 	
@@ -26,7 +26,7 @@ void Timer_Mode_OCPWM_Config(TIM_HandleTypeDef * Timer, unsigned int mode, unsig
 	TIM_PWM_OC.OCFastMode = TIM_OCFAST_DISABLE;
 }
 
-void Timer_PWM_OCPWM_Pulse_Channel(TIM_HandleTypeDef * Timer, unsigned int channel, unsigned int pulse){
+void Timer_PWM_OCPWM_Pulse_Channel(TIM_HandleTypeDef * Timer, uint16_t channel, uint16_t pulse){
 	HAL_TIM_OC_Init(Timer);
 	TIM_OC_InitTypeDef TIM_PWM_OC;
 	
@@ -34,7 +34,7 @@ void Timer_PWM_OCPWM_Pulse_Channel(TIM_HandleTypeDef * Timer, unsigned int chann
 	HAL_TIM_OC_ConfigChannel(Timer,&TIM_PWM_OC, channel);
 }
 
-void Timer_Mode_IC_Config(TIM_HandleTypeDef * Timer, unsigned int polarity, unsigned int filter, unsigned int selection, unsigned int prescaler){
+void Timer_Mode_IC_Config(TIM_HandleTypeDef * Timer, uint16_t polarity, uint16_t filter, uint16_t selection, uint16_t prescaler){
 	HAL_TIM_IC_Init(Timer);
 	TIM_IC_InitTypeDef TIM_IC;
 	
@@ -79,7 +79,7 @@ void ENABLE_CLK_TIMER_CHOICE(TIM_TypeDef * timer){
 /*									GPIO Configuration															*/
 /*--------------------------------------------------------------------------------------------------------------*/
 
-void GPIO_Configuration(GPIO_TypeDef * GPIO, unsigned int mode, unsigned int Pin, unsigned int alternate){
+void GPIO_Configuration(GPIO_TypeDef * GPIO, uint16_t mode, uint16_t Pin, uint16_t alternate){
 	
 	ENABLE_CLK_GPIO(GPIO);
